Added grind rail and continued developing leve
I need to look at what happens when you miss the grind arrow, as at the moment you just go through the rail.
I think I might get it to make you swerve around it. But from there I aim to add more obstacles and things to the map to add interest to the movement and tricks. But in terms of core controls, its done!